Fixed pylint findings
