+++ webots/worlds/GridWithMarkers.wbt
@@ -42,7 +42,7 @@ DEF ROBOT Zumo32U4 {
 }
 Supervisor {
   name "Supervisor"
-  controller "Supervisor"
+  controller "supervisor"
   supervisor TRUE
 }
 Marker {
