+++ webots/worlds/zumo_with_com_system/LineFollowerTrack.wbt
@@ -45,6 +45,6 @@ DEF ROBOT Zumo32U4 {
 }
 Supervisor {
   name "Supervisor"
-  controller "Supervisor"
+  controller "supervisor"
   supervisor TRUE
 }
